feat: expose useDefaults() (#16461)
