+++ packages/vuetify/src/composables/defaults.ts
@@ -1,9 +1,12 @@
+// Composables
+import { useToggleScope } from '@/composables/toggleScope'
+
 // Utilities
-import { computed, inject, provide, ref, unref } from 'vue'
-import { mergeDeep } from '@/util'
+import { computed, inject, provide, ref, shallowRef, unref, watchEffect } from 'vue'
+import { getCurrentInstance, injectSelf, mergeDeep, toKebabCase } from '@/util'
 
 // Types
-import type { ComputedRef, InjectionKey, Ref } from 'vue'
+import type { ComputedRef, InjectionKey, Ref, VNode } from 'vue'
 import type { MaybeRef } from '@/util'
 
 export type DefaultsInstance = undefined | {
@@ -19,7 +22,7 @@ export function createDefaults (options?: DefaultsInstance): Ref<DefaultsInstanc
   return ref(options)
 }
 
-export function useDefaults () {
+export function injectDefaults () {
   const defaults = inject(DefaultsSymbol)
 
   if (!defaults) throw new Error('[Vuetify] Could not find defaults instance')
@@ -36,7 +39,7 @@ export function provideDefaults (
     scoped?: MaybeRef<boolean | undefined>
   }
 ) {
-  const injectedDefaults = useDefaults()
+  const injectedDefaults = injectDefaults()
   const providedDefaults = ref(defaults)
 
   const newDefaults = computed(() => {
@@ -74,4 +77,53 @@ export function provideDefaults (
   provide(DefaultsSymbol, newDefaults)
 
   return newDefaults
+}
+
+function propIsDefined (vnode: VNode, prop: string) {
+  return typeof vnode.props?.[prop] !== 'undefined' ||
+    typeof vnode.props?.[toKebabCase(prop)] !== 'undefined'
+}
+
+export function useDefaults (props: Record<string, any>, name?: string, defaults = injectDefaults()) {
+  const vm = getCurrentInstance('useDefaults')
+
+  name = name ?? vm.type.name ?? vm.type.__name
+  if (!name) {
+    throw new Error('[Vuetify] Could not determine component name')
+  }
+
+  const componentDefaults = computed(() => defaults.value![props._as ?? name])
+  const _props = new Proxy(props, {
+    get (target, prop) {
+      const propValue = Reflect.get(target, prop)
+      if (prop === 'class' || prop === 'style') {
+        return [componentDefaults.value?.[prop], propValue].filter(v => v != null)
+      } else if (typeof prop === 'string' && !propIsDefined(vm.vnode, prop)) {
+        return componentDefaults.value?.[prop] ?? defaults.value!.global?.[prop] ?? propValue
+      }
+      return propValue
+    },
+  })
+
+  const _subcomponentDefaults = shallowRef()
+  watchEffect(() => {
+    if (componentDefaults.value) {
+      const subComponents = Object.entries(componentDefaults.value).filter(([key]) => key.startsWith(key[0].toUpperCase()))
+      if (subComponents.length) _subcomponentDefaults.value = Object.fromEntries(subComponents)
+    }
+  })
+
+  function provideSubDefaults () {
+    // If subcomponent defaults are provided, override any
+    // subcomponents provided by the component's setup function.
+    // This uses injectSelf so must be done after the original setup to work.
+    useToggleScope(_subcomponentDefaults, () => {
+      provideDefaults(mergeDeep(
+        injectSelf(DefaultsSymbol)?.value ?? {},
+        _subcomponentDefaults.value
+      ))
+    })
+  }
+
+  return { props: _props, provideSubDefaults }
 }
