+++ packages/vuetify/src/util/defineComponent.tsx
@@ -1,17 +1,11 @@
 // Utils
 import {
   defineComponent as _defineComponent, // eslint-disable-line no-restricted-imports
-  computed,
-  getCurrentInstance,
-  shallowRef,
-  watchEffect,
 } from 'vue'
 import { consoleWarn } from '@/util/console'
-import { mergeDeep, pick, toKebabCase } from '@/util/helpers'
-import { injectSelf } from '@/util/injectSelf'
+import { pick, toKebabCase } from '@/util/helpers'
 import { propsFactory } from '@/util/propsFactory'
-import { DefaultsSymbol, provideDefaults, useDefaults } from '@/composables/defaults'
-import { useToggleScope } from '@/composables/toggleScope'
+import { injectDefaults, useDefaults } from '@/composables/defaults'
 
 // Types
 import type {
@@ -31,15 +25,9 @@ import type {
   FunctionalComponent,
   MethodOptions,
   ObjectEmitsOptions,
-  VNode,
   VNodeChild,
   VNodeProps,
 } from 'vue'
-
-function propIsDefined (vnode: VNode, prop: string) {
-  return typeof vnode.props?.[prop] !== 'undefined' ||
-    typeof vnode.props?.[toKebabCase(prop)] !== 'undefined'
-}
 
 // No props
 export function defineComponent<
@@ -118,44 +106,16 @@ export function defineComponent (options: ComponentOptions) {
 
     options.props._as = String
     options.setup = function setup (props: Record<string, any>, ctx) {
-      const defaults = useDefaults()
+      const defaults = injectDefaults()
 
       // Skip props proxy if defaults are not provided
       if (!defaults.value) return options._setup(props, ctx)
 
-      const vm = getCurrentInstance()!
-      const componentDefaults = computed(() => defaults.value![props._as ?? options.name!])
-      const _props = new Proxy(props, {
-        get (target, prop) {
-          const propValue = Reflect.get(target, prop)
-          if (prop === 'class' || prop === 'style') {
-            return [componentDefaults.value?.[prop], propValue].filter(v => v != null)
-          } else if (typeof prop === 'string' && !propIsDefined(vm.vnode, prop)) {
-            return componentDefaults.value?.[prop] ?? defaults.value!.global?.[prop] ?? propValue
-          }
-          return propValue
-        },
-      })
-
-      const _subcomponentDefaults = shallowRef()
-      watchEffect(() => {
-        if (componentDefaults.value) {
-          const subComponents = Object.entries(componentDefaults.value).filter(([key]) => key.startsWith(key[0].toUpperCase()))
-          if (subComponents.length) _subcomponentDefaults.value = Object.fromEntries(subComponents)
-        }
-      })
+      const { props: _props, provideSubDefaults } = useDefaults(props, props._as ?? options.name, defaults)
 
       const setupBindings = options._setup(_props, ctx)
 
-      // If subcomponent defaults are provided, override any
-      // subcomponents provided by the component's setup function.
-      // This uses injectSelf so must be done after the original setup to work.
-      useToggleScope(_subcomponentDefaults, () => {
-        provideDefaults(mergeDeep(
-          injectSelf(DefaultsSymbol)?.value ?? {},
-          _subcomponentDefaults.value
-        ))
-      })
+      provideSubDefaults()
 
       return setupBindings
     }
