+++ packages/vuetify/src/util/index.ts
@@ -10,6 +10,7 @@ export * from './getCurrentInstance'
 export * from './getScrollParent'
 export * from './globals'
 export * from './helpers'
+export * from './injectSelf'
 export * from './isFixedPosition'
 export * from './propsFactory'
 export * from './useRender'
